+++ src/Sensorario/Validators/Validator.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Sensorario\Validators;
+
+use Sensorario\ValueObject\ValueObject;
+
+interface Validator
+{
+    public static function check(ValueObject $valueObject);
+}